requirements.txt
numpy
bayesian-optimization
scikit-image
pytest
fastapi
uvicorn
matplotlib
httpx
strawberry-graphql[fastapi]
psycopg2-binary
